:sparkles: move other related files from app to inside module in Modules\Advertise
