+++ app/Http/Controllers/Admin/Advertise/AdvertisementController.php
@@ -11,10 +11,10 @@ use App\Http\Resources\Admin\Advertise\AdvertisementCollection;
 use App\Http\Resources\Admin\Advertise\AdvertisementResource;
 use App\Http\Responses\ApiJsonResponse;
 use App\Http\Services\Image\ImageService;
-use App\Jobs\ProcessNewAdvertisementJob;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Http\Resources\Json\ResourceCollection;
+use Modules\Advertise\Jobs\ProcessNewAdvertisementJob;
 use Modules\Advertise\Models\Advertisement;
 
 final class AdvertisementController extends Controller
