+++ modules/advertise/src/Notifications/NewAdPostedNotification.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace App\Notifications;
+declare(strict_types=1);
+
+namespace Modules\Advertise\Notifications;
 
 use App\Enums\Queue;
 use Illuminate\Bus\Queueable;
@@ -8,7 +10,7 @@ use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\Messages\MailMessage;
 use Illuminate\Notifications\Notification;
 
-class NewAdPostedNotification extends Notification implements ShouldQueue
+final class NewAdPostedNotification extends Notification implements ShouldQueue
 {
     use Queueable;
 
@@ -25,7 +27,7 @@ class NewAdPostedNotification extends Notification implements ShouldQueue
 
     public function toMail(object $notifiable): MailMessage
     {
-        return (new MailMessage)
+        return (new MailMessage())
             ->line('The introduction to the notification.')
             ->action('Notification Action', url('/'))
             ->line('Thank you for using our application!');
@@ -34,7 +36,7 @@ class NewAdPostedNotification extends Notification implements ShouldQueue
     public function toArray(object $notifiable): array
     {
         return [
-            //
+
         ];
     }
 }
